+++ src/util/date.test.ts
@@ -1,8 +1,9 @@
 import { getVmScanSinceDate } from './date';
 
 describe('#getVmScanSinceDate', () => {
-  afterEach(() => {
-    process.env.VM_SCAN_SINCE_DAYS = undefined;
+  beforeEach(() => {
+    delete process.env.VM_SCAN_SINCE_DAYS;
+    delete process.env.VM_SCAN_SINCE_MS;
     jest.resetAllMocks();
   });
 
@@ -11,9 +12,15 @@ describe('#getVmScanSinceDate', () => {
     expect(getVmScanSinceDate()).toEqual('2020-09-12T18:43:44Z');
   });
 
-  test('should allow configuring range using environment variable', () => {
+  test('should allow configuring range in days using environment variable', () => {
     jest.spyOn(Date, 'now').mockImplementationOnce(() => 1602528224084);
     process.env.VM_SCAN_SINCE_DAYS = '7';
     expect(getVmScanSinceDate()).toEqual('2020-10-05T18:43:44Z');
+  });
+
+  test('should allow configuring range in milliseconds using environment variable', () => {
+    jest.spyOn(Date, 'now').mockImplementationOnce(() => 1602528224084);
+    process.env.VM_SCAN_SINCE_MS = '3600000';
+    expect(getVmScanSinceDate()).toEqual('2020-10-12T17:43:44Z');
   });
 });